+++ Sources/JSON/JSONEncodable.swift
@@ -1,0 +1,1 @@
+public protocol JSONEncodable: Encodable {}
